Applied patch 1206278 from Mike Wetherell. Changes ceilf calls to ceil to be compatible with Mac OS 10.1.
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@34281 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/cocoa/button.mm
@@ -106,6 +106,6 @@ wxSize wxButtonBase::GetDefaultSize()
 {
     static NSRect cocoaRect = MakeNSButtonDefaultRect();
     // Apple HIG says OK/Cancel buttons have default width of 68.
-    return wxSize(68,(int)ceilf(cocoaRect.size.height));
+    return wxSize(68,(int)ceil(cocoaRect.size.height));
 }
