+++ src/cocoa/control.mm
@@ -87,7 +87,7 @@ NS_ENDHANDLER
     if(cell)
     {
         NSSize cellSize = [cell cellSize];
-        wxSize size((int)ceilf(cellSize.width),(int)ceilf(cellSize.height));
+        wxSize size((int)ceil(cellSize.width),(int)ceil(cellSize.height));
         wxLogTrace(wxTRACE_COCOA_Window_Size,wxT("wxControl=%p::DoGetBestSize()==(%d,%d) from NSCell"),this,size.x,size.y);
         return size;
     }
@@ -104,7 +104,7 @@ NS_ENDHANDLER
     if(didFit)
     {
         NSRect cocoaRect = [m_cocoaNSView frame];
-        wxSize size((int)ceilf(cocoaRect.size.width),(int)ceilf(cocoaRect.size.height));
+        wxSize size((int)ceil(cocoaRect.size.width),(int)ceil(cocoaRect.size.height));
         [m_cocoaNSView setFrame: storedRect];
         wxLogTrace(wxTRACE_COCOA_Window_Size,wxT("wxControl=%p::DoGetBestSize()==(%d,%d) from sizeToFit"),this,size.x,size.y);
         return size;
